Flatten Fabric-specific code into platform package
