+++ src/main/java/io/github/gaming32/bingo/client/recipeviewer/RecipeViewerPlugin.java
@@ -7,10 +7,10 @@ import net.minecraft.world.item.ItemStack;
 
 public abstract class RecipeViewerPlugin {
     public static RecipeViewerPlugin detect() {
-        if (BingoPlatform.platform.isModLoaded("jei")) {
+        if (BingoPlatform.isModLoaded("jei")) {
             return new JEIPlugin();
         }
-//        if (BingoPlatform.platform.isModLoaded("roughlyenoughitems")) {
+//        if (BingoPlatform.isModLoaded("roughlyenoughitems")) {
 //            return new REIPlugin();
 //        }
         return new NoPlugin();
