+++ src/main/java/io/github/gaming32/bingo/network/BingoNetworking.java
@@ -1,8 +1,14 @@
 package io.github.gaming32.bingo.network;
 
 import io.github.gaming32.bingo.platform.BingoPlatform;
+import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationNetworking;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
+import net.fabricmc.fabric.api.networking.v1.ServerConfigurationNetworking;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.client.multiplayer.ClientCommonPacketListenerImpl;
 import net.minecraft.network.ConnectionProtocol;
+import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.PacketListener;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
@@ -21,46 +27,92 @@ import org.jetbrains.annotations.Nullable;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
-public abstract class BingoNetworking {
+public final class BingoNetworking {
     public static final int PROTOCOL_VERSION = 12;
 
-    public static BingoNetworking instance() {
-        return BingoPlatform.platform.getNetworking();
+    public static void onRegister(Consumer<Registrar> handler) {
+        handler.accept(new Registrar());
     }
 
-    public abstract void onRegister(Consumer<Registrar> handler);
+    public static void sendToServer(CustomPacketPayload payload) {
+        ClientPlayNetworking.send(payload);
+    }
 
-    public abstract void sendToServer(CustomPacketPayload payload);
+    public static void sendTo(ServerPlayer player, CustomPacketPayload payload) {
+        ServerPlayNetworking.send(player, payload);
+    }
 
-    public abstract void sendTo(ServerPlayer player, CustomPacketPayload payload);
-
-    public void sendTo(Iterable<ServerPlayer> players, CustomPacketPayload payload) {
+    public static void sendTo(Iterable<ServerPlayer> players, CustomPacketPayload payload) {
         for (final ServerPlayer player : players) {
             sendTo(player, payload);
         }
     }
 
-    public abstract boolean canServerReceive(CustomPacketPayload.Type<?> type);
+    public static boolean canServerReceive(CustomPacketPayload.Type<?> type) {
+        return ClientPlayNetworking.canSend(type);
+    }
 
-    public abstract boolean canPlayerReceive(ServerPlayer player, CustomPacketPayload.Type<?> type);
+    public static boolean canPlayerReceive(ServerPlayer player, CustomPacketPayload.Type<?> type) {
+        return ServerPlayNetworking.canSend(player, type);
+    }
 
-    protected abstract void finishTask(ServerConfigurationPacketListenerImpl packetListener, ConfigurationTask.Type type);
-
-    public final void finishTask(Context context, ConfigurationTask.Type type) {
+    public static void finishTask(Context context, ConfigurationTask.Type type) {
         if (!(context.packetListener instanceof ServerConfigurationPacketListenerImpl packetListener)) {
             throw new IllegalStateException("finishTask can only be called during the configuration phase");
         }
-        finishTask(packetListener, type);
+        packetListener.completeTask(type);
     }
 
-    public abstract static class Registrar {
-        public abstract <P extends CustomPacketPayload> void register(
+    public static final class Registrar {
+        private Registrar() {
+        }
+
+        @SuppressWarnings("unchecked")
+        public <P extends CustomPacketPayload> void register(
             ConnectionProtocol protocol,
             @Nullable PacketFlow flow,
             CustomPacketPayload.Type<P> type,
             StreamCodec<? super RegistryFriendlyByteBuf, P> codec,
             BiConsumer<P, Context> handler
-        );
+        ) {
+            if (flow == null || flow == PacketFlow.CLIENTBOUND) {
+                switch (protocol) {
+                    case PLAY -> PayloadTypeRegistry.clientboundPlay().register(type, codec);
+                    case CONFIGURATION -> PayloadTypeRegistry.clientboundConfiguration().register(type, (StreamCodec<? super FriendlyByteBuf, P>) codec);
+                    default -> throw new IllegalArgumentException("Cannot register for connection state: " + protocol);
+                }
+                if (BingoPlatform.isClient()) {
+                    ClientReceiverRegistrar.register(protocol, type, handler);
+                }
+            }
+            if (flow == null || flow == PacketFlow.SERVERBOUND) {
+                switch (protocol) {
+                    case PLAY -> {
+                        PayloadTypeRegistry.serverboundPlay().register(type, codec);
+                        ServerPlayNetworking.registerGlobalReceiver(type, (payload, context) ->
+                            handler.accept(payload, new Context(
+                                context.player(),
+                                context.responseSender()::sendPacket,
+                                context.player().connection,
+                                PacketFlow.SERVERBOUND
+                            ))
+                        );
+                    }
+                    case CONFIGURATION -> {
+                        PayloadTypeRegistry.serverboundConfiguration().register(type, (StreamCodec<? super FriendlyByteBuf, P>) codec);
+                        ServerConfigurationNetworking.registerGlobalReceiver(type, (payload, context) ->
+                            handler.accept(payload, new Context(
+                                null,
+                                context.responseSender()::sendPacket,
+                                context.packetListener(),
+                                PacketFlow.SERVERBOUND
+                            ))
+                        );
+                    }
+                    default -> throw new IllegalArgumentException("Cannot register for connection state: " + protocol);
+                }
+            }
+        }
 
         public <P extends AbstractCustomPayload> void register(
             @Nullable PacketFlow flow,
@@ -68,6 +120,33 @@ public abstract class BingoNetworking {
             StreamCodec<? super RegistryFriendlyByteBuf, P> codec
         ) {
             register(ConnectionProtocol.PLAY, flow, type, codec, AbstractCustomPayload::handle);
+        }
+
+        private static class ClientReceiverRegistrar {
+            public static <P extends CustomPacketPayload> void register(
+                ConnectionProtocol protocol, CustomPacketPayload.Type<P> type, BiConsumer<P, Context> handler
+            ) {
+                switch (protocol) {
+                    case PLAY -> ClientPlayNetworking.registerGlobalReceiver(type, (payload, context) ->
+                        handler.accept(payload, new Context(
+                            context.player(),
+                            context.responseSender()::sendPacket,
+                            context.player().connection,
+                            PacketFlow.CLIENTBOUND
+                        ))
+                    );
+                    case CONFIGURATION -> ClientConfigurationNetworking.registerGlobalReceiver(type, (payload, context) ->
+                        handler.accept(payload, new Context(
+                            null,
+                            context.responseSender()::sendPacket,
+                            context.packetListener(),
+                            PacketFlow.CLIENTBOUND
+                        ))
+                    );
+                    default -> throw new IllegalArgumentException("Cannot register for connection state: " + protocol);
+                }
+
+            }
         }
     }
 
@@ -108,7 +187,7 @@ public abstract class BingoNetworking {
         public void disconnect(Component reason) {
             if (packetListener instanceof ServerCommonPacketListenerImpl serverListener) {
                 serverListener.disconnect(reason);
-            } else if (!BingoPlatform.platform.isClient() || !ClientDisconnecter.disconnect(packetListener, reason)) {
+            } else if (!BingoPlatform.isClient() || !ClientDisconnecter.disconnect(packetListener, reason)) {
                 throw new IllegalStateException("Cannot disconnect with listener " + packetListener.getClass().getName());
             }
         }
