+++ src/main/java/io/github/gaming32/bingo/platform/registry/DeferredRegister.java
@@ -4,20 +4,33 @@ import io.github.gaming32.bingo.util.Identifiers;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.Identifier;
-import org.jetbrains.annotations.ApiStatus;
 
 import java.util.function.Supplier;
 
-public interface DeferredRegister<T> {
-    Registry<T> registry();
+public final class DeferredRegister<T> {
+    private final Registry<T> registry;
 
-    ResourceKey<? extends Registry<T>> registryKey();
+    private DeferredRegister(Registry<T> registry) {
+        this.registry = registry;
+    }
 
-    @ApiStatus.Internal
-    <S extends T> RegistryValue<S> register(Identifier location, Supplier<S> value);
+    public static <T> DeferredRegister<T> create(Registry<T> registry) {
+        return new DeferredRegister<>(registry);
+    }
 
-    @ApiStatus.Internal
-    default <S extends T> RegistryValue<S> register(String id, Supplier<S> value) {
+    public Registry<T> registry() {
+        return registry;
+    }
+
+    public ResourceKey<? extends Registry<T>> registryKey() {
+        return registry.key();
+    }
+
+    public <S extends T> RegistryValue<S> register(Identifier location, Supplier<S> value) {
+        return new RegistryValue<>(Registry.registerForHolder(registry, location, value.get()));
+    }
+
+    public <S extends T> RegistryValue<S> register(String id, Supplier<S> value) {
         return register(Identifiers.bingo(id), value);
     }
 }
