+++ src/main/java/io/github/gaming32/bingo/platform/registry/RegistryBuilder.java
@@ -1,7 +1,8 @@
 package io.github.gaming32.bingo.platform.registry;
 
-import io.github.gaming32.bingo.platform.BingoPlatform;
 import io.github.gaming32.bingo.util.Identifiers;
+import net.fabricmc.fabric.api.event.registry.FabricRegistryBuilder;
+import net.fabricmc.fabric.api.event.registry.RegistryAttribute;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.Identifier;
@@ -30,18 +31,16 @@ public class RegistryBuilder<T> {
     }
 
     public DeferredRegister<T> build() {
-        return BingoPlatform.platform.buildDeferredRegister(this);
-    }
+        final var fabricBuilder = defaultId != null
+            ? FabricRegistryBuilder.createDefaulted(key, defaultId)
+            : FabricRegistryBuilder.create(key);
+        if (synced) {
+            fabricBuilder.attribute(RegistryAttribute.SYNCED);
+        }
 
-    public ResourceKey<Registry<T>> getKey() {
-        return key;
-    }
+        // FIXME: this technically shouldn't be optional, but prevents clients from being kicked on different mod loaders
+        fabricBuilder.attribute(RegistryAttribute.OPTIONAL);
 
-    public boolean isSynced() {
-        return synced;
-    }
-
-    public Identifier getDefaultId() {
-        return defaultId;
+        return DeferredRegister.create(fabricBuilder.buildAndRegister());
     }
 }
