+++ src/main/java/io/github/gaming32/bingo/subpredicates/BingoEntitySubPredicates.java
@@ -1,7 +1,6 @@
 package io.github.gaming32.bingo.subpredicates;
 
 import com.mojang.serialization.MapCodec;
-import io.github.gaming32.bingo.platform.BingoPlatform;
 import io.github.gaming32.bingo.platform.registry.DeferredRegister;
 import io.github.gaming32.bingo.platform.registry.RegistryValue;
 import io.github.gaming32.bingo.util.Identifiers;
@@ -10,7 +9,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 
 public class BingoEntitySubPredicates {
     private static final DeferredRegister<MapCodec<? extends EntitySubPredicate>> REGISTER =
-        BingoPlatform.platform.createDeferredRegister(BuiltInRegistries.ENTITY_SUB_PREDICATE_TYPE);
+        DeferredRegister.create(BuiltInRegistries.ENTITY_SUB_PREDICATE_TYPE);
 
     public static final RegistryValue<MapCodec<BingoPlayerPredicate>> PLAYER = register("player", BingoPlayerPredicate.CODEC);
     public static final RegistryValue<MapCodec<ItemEntityPredicate>> ITEM = register("item", ItemEntityPredicate.CODEC);
