+++ src/main/resources/bingo.mixins.json
@@ -41,7 +41,6 @@
     "MixinOcelot",
     "MixinPig",
     "MixinPlayerAdvancements",
-    "MixinPlayerList",
     "MixinRangedCrossbowAttackGoal",
     "MixinRavager",
     "MixinRespawnAnchorBlock",
